refactor: separate eth and gateway requests
