+++ tools/walletextension/frontend/src/components/providers/wallet-provider.tsx
@@ -1,25 +1,23 @@
 import { createContext, useContext, useEffect, useState } from "react";
-import { ethers } from "ethers";
 import {
   WalletConnectionContextType,
   WalletConnectionProviderProps,
   Account,
 } from "../../types/interfaces/WalletInterfaces";
 import { showToast } from "../ui/use-toast";
-import {
-  getRandomIntAsString,
-  isTenChain,
-  isValidUserIDFormat,
-} from "../../lib/utils";
+import { isValidUserIDFormat } from "../../lib/utils";
 import {
   accountIsAuthenticated,
-  authenticateAccountWithTenGatewayEIP712,
   fetchVersion,
   revokeAccountsApi,
 } from "../../api/gateway";
-import { METAMASK_CONNECTION_TIMEOUT } from "../../lib/constants";
-import { requestMethods } from "@/routes";
 import { ToastType } from "@/types/interfaces";
+import {
+  authenticateAccountWithTenGatewayEIP712,
+  getUserID,
+} from "@/api/ethRequests";
+import { ethers } from "ethers";
+import ethService from "@/services/ethService";
 
 const { ethereum } = typeof window !== "undefined" ? window : ({} as any);
 
@@ -28,6 +26,7 @@ const WalletConnectionContext =
 
 export const useWalletConnection = (): WalletConnectionContextType => {
   const context = useContext(WalletConnectionContext);
+
   if (!context) {
     throw new Error(
       "useWalletConnection must be used within a WalletConnectionProvider"
@@ -52,7 +51,9 @@ export const WalletConnectionProvider = ({
         return;
       }
       if (userID && isValidUserIDFormat(userID)) {
-        await displayCorrectScreenBasedOnMetamaskAndUserID(userID, provider);
+        const status =
+          await ethService.getCorrectScreenBasedOnMetamaskAndUserID(userID);
+        setWalletConnected(status);
       }
     };
     ethereum.on("accountsChanged", handleAccountsChanged);
@@ -63,88 +64,26 @@ export const WalletConnectionProvider = ({
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
+  const initialize = async () => {
+    const providerInstance = new ethers.providers.Web3Provider(ethereum);
+    setProvider(providerInstance);
+    await ethService.checkIfMetamaskIsLoaded(providerInstance);
+    const id = await getUserID(providerInstance);
+    setUserID(id);
+    const status = await ethService.getCorrectScreenBasedOnMetamaskAndUserID(
+      id
+    );
+    setWalletConnected(status);
+    const accounts = await ethService.getAccounts(providerInstance);
+    setAccounts(accounts || null);
+    setVersion(await fetchVersion());
+    setLoading(false);
+  };
+
   useEffect(() => {
-    const initialize = async () => {
-      await checkIfMetamaskIsLoaded();
-    };
     initialize();
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
-
-  const checkIfMetamaskIsLoaded = async () => {
-    if (ethereum) {
-      await handleEthereum();
-    } else {
-      showToast(ToastType.INFO, "Connecting to MetaMask...");
-      window.addEventListener("ethereum#initialized", handleEthereum, {
-        once: true,
-      });
-      setTimeout(handleEthereum, METAMASK_CONNECTION_TIMEOUT);
-    }
-  };
-
-  const handleEthereum = async () => {
-    if (ethereum && ethereum.isMetaMask) {
-      const provider = new ethers.providers.Web3Provider(ethereum);
-      setProvider(provider);
-      const fetchedUserID = await getUserID(provider);
-      await displayCorrectScreenBasedOnMetamaskAndUserID(
-        fetchedUserID,
-        provider
-      );
-    } else {
-      showToast(
-        ToastType.WARNING,
-        "Please install MetaMask to use Ten Gateway."
-      );
-    }
-  };
-
-  const getUserID = async (provider: ethers.providers.Web3Provider) => {
-    if (!provider) {
-      return null;
-    }
-
-    try {
-      if (await isTenChain()) {
-        const id = await provider.send(requestMethods.getStorageAt, [
-          "getUserID",
-          getRandomIntAsString(0, 1000),
-          null,
-        ]);
-        setUserID(id);
-        return id;
-      } else {
-        return null;
-      }
-    } catch (e: any) {
-      showToast(
-        ToastType.DESTRUCTIVE,
-        `${e.message} ${e.data?.message}` ||
-          "Error: Could not fetch your user ID. Please try again later."
-      );
-      console.error(e);
-      return null;
-    }
-  };
-
-  const displayCorrectScreenBasedOnMetamaskAndUserID = async (
-    userID: string,
-    provider: ethers.providers.Web3Provider
-  ) => {
-    setVersion(await fetchVersion());
-    if (await isTenChain()) {
-      if (userID) {
-        await getAccounts(provider, userID);
-      } else {
-        setWalletConnected(false);
-      }
-    } else {
-      setWalletConnected(false);
-    }
-
-    setLoading(false);
-  };
 
   const connectAccount = async (account: string) => {
     if (!userID) {
@@ -182,69 +121,27 @@ export const WalletConnectionProvider = ({
       showToast(ToastType.DESTRUCTIVE, "Accounts revoked!");
       setAccounts(null);
       setWalletConnected(false);
+      setUserID("");
     }
   };
 
-  const getAccounts = async (
-    provider: ethers.providers.Web3Provider,
-    id: string
-  ) => {
-    try {
-      if (!provider) {
-        showToast(
-          ToastType.DESTRUCTIVE,
-          "No provider found. Please try again later."
-        );
-        return;
-      }
-
-      showToast(ToastType.INFO, "Getting accounts...");
-
-      if (!(await isTenChain())) {
-        showToast(ToastType.DESTRUCTIVE, "Please connect to the Ten chain.");
-        return;
-      }
-
-      const accounts = await provider.listAccounts();
-
-      if (accounts.length === 0) {
-        showToast(ToastType.DESTRUCTIVE, "No MetaMask accounts found.");
-        return;
-      }
-
-      let updatedAccounts: Account[] = [];
-
-      for (let i = 0; i < accounts.length; i++) {
-        const account = accounts[i];
-        authenticateAccountWithTenGatewayEIP712(id, account);
-        const { status } = await accountIsAuthenticated(id, account);
-        updatedAccounts.push({
-          name: account,
-          connected: status,
-        });
-      }
-
-      setAccounts(updatedAccounts);
-      setWalletConnected(true);
-
-      showToast(ToastType.SUCCESS, "Accounts authenticated");
-    } catch (error) {
-      console.error(error);
-      showToast(ToastType.DESTRUCTIVE, "An error occurred. Please try again.");
-    }
+  const fetchUserAccounts = async () => {
+    const accounts = await ethService.getAccounts(provider);
+    setAccounts(accounts || null);
+    setWalletConnected(true);
   };
 
   const walletConnectionContextValue: WalletConnectionContextType = {
     walletConnected,
     accounts,
     userID,
-    setUserID,
     connectAccount,
-    provider,
     version,
     revokeAccounts,
-    getAccounts,
     loading,
+    provider,
+    fetchUserAccounts,
+    setLoading,
   };
 
   return (
